+++ ricaun.RevitTest.Console/Revit/Utils/RevitTestUtils.cs
@@ -55,11 +55,16 @@ namespace ricaun.RevitTest.Console.Revit.Utils
                 }
             }
 
+            if (tests.Length == 0)
+            {
+                throw new Exception($"Unable to read tests class using the Revit version {revitVersion}.");
+            }
+
 #if DEBUG
             LoggerTest($"Test °C");
             LoggerTest($"Length {tests.Length}");
             LoggerTest($"TestEngine {TestEngine.Version.ToString(3)}");
-            LoggerTest($"Info {AppUtils.GetInfo()}");
+            LoggerTest($"{AppUtils.GetInfo()}");
             if (LoggerTests.Any())
             {
                 LoggerTests.AddRange(tests);